Fixed DietCheck skipping if logged in.
