+++ src/App.js
@@ -85,7 +85,7 @@ function App() {
         <Route path="/keywordGrab"element={<KeywordGrab/>}/>
         <Route path="/priceCheck"element={<PriceGrab/>}/>
         <Route path="/timeGrab"element={<TimeGrab/>}/>
-        <Route path="/dietaryRestrictions"element={<DietCheck/>}/>
+        <Route path="/dietaryRestrictions"element={<DietCheck user={user}/>}/>
         <Route path="/location"element={<DistanceGrab user={user}/>}/>
         <Route path="/account/filters" element={<Preferences user={user}/>}/>
         <Route path="/account/allergies" element={<Allergies user={user}/>}/>
